Moved async_manual_reset_event to v1/, backward compatible
