+++ source/CMakeLists.txt
@@ -12,7 +12,7 @@ target_sources(unifex
   PRIVATE
     atomic_intrusive_list.cpp
     async_auto_reset_event.cpp
-    async_manual_reset_event.cpp
+    async_manual_reset_event_v1.cpp
     async_mutex_v1.cpp
     async_mutex_v2.cpp
     async_pass.cpp
