+++ test/CMakeLists.txt
@@ -15,6 +15,6 @@ foreach(file-path ${test-sources})
 endforeach()
 
 target_link_libraries(any_sender_of_test PUBLIC GTest::gmock_main)
-target_link_libraries(async_manual_reset_event_test PUBLIC GTest::gmock_main)
+target_link_libraries(async_manual_reset_event_v1_test PUBLIC GTest::gmock_main)
 target_link_libraries(async_scope_test PUBLIC GTest::gmock_main)
 target_link_libraries(async_scope_v0_test PUBLIC GTest::gmock_main)
